* Fixed: In loadcle and a member of the editor I assumed a char variable can hold a negative value or at least the compiler is smart enough to assume it means 0xFF. It isn't true for Power PC platforms. I changed it to work with explicit unsigned chars. * Fixed: The edspecs module used a non-endian clean code when saving the environment strings producing SIGSEGVs when loading the file in big endian systems. Lamentably I think it there are more things like it. Seen in Power PC.

+++ setedit/mainsrc/edspecs.cc
@@ -168,7 +168,7 @@ void fPutVar(char *var,FILE *f)
  char *c=(char *)GetVariable(var);
  if (c)
    {
-    int len=strlen(var);
+    ushort len=strlen(var);
     fwrite(&len,sizeof(ushort),1,f);
     fwrite(var,len,1,f);
     len=strlen(c);
